Minor improvements - Add missing NonNull annotations for java - Change zIndex type from double to int for groundOverlays to avoid future issues on iOS platform - Documentation improvements

+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/Convert.java
@@ -932,7 +932,7 @@ class Convert {
             .setHeight((double) groundOverlay.getWidth())
             .setBearing((double) groundOverlay.getBearing())
             .setTransparency((double) groundOverlay.getTransparency())
-            .setZIndex((double) groundOverlay.getZIndex())
+            .setZIndex((long) groundOverlay.getZIndex())
             .setVisible(groundOverlay.isVisible())
             .setClickable(groundOverlay.isClickable());
